requirements.txt
requests
py-cord
python-dotenv
Pillow
easy-pil
chat-exporter
aiosqlite
pyotp
aiohttp